requirements.txt

viam-sdk[mlmodel]==0.45.0

typing-extensions

ncnn
